refactor(combat): split Quick Battle engine into concern libraries (#4545)
Extract group/CP bookkeeping, strike-strength math, and emplaced mixed-loss
helpers from quick_battle_resolver_engine.dart into sibling libraries. The
engine entry re-exports all symbols so resolver import paths stay stable (Slice B).
