+++ SPEC/program/repo-lint.md
@@ -102,7 +102,7 @@
 | `tool/check_map_region_dispatch_central.dart` | Keep old/new-world region *dispatch* centralized in `packages/colonizethis_map/lib/**`: forbid any inline `==` / `!=` comparison against `kRegionOldWorld` or `kRegionNewWorld` outside the canonical dispatch file `packages/colonizethis_map/lib/src/map_region_dispatch.dart`. Colour, capital-scope, port-tile bucketing, and view-data assembly paths must resolve old/new branching via `selectByMapRegionId(...)` / `selectByMapRegionIdOrNull(...)` so dispatch has one source of truth. Using the constants as map keys, list elements, named arguments, or string-interpolation operands is unaffected; the canonical file is exempt and comment lines are ignored (Refs #3574); rule `repo.map_region_dispatch_central` |
 | `tool/check_map_gen_no_image_import.dart` | Enforce the `colonizethis_map` generation→render layer boundary in `packages/colonizethis_map/lib/**`: forbid any `import` / `export` of `package:image` outside the render-layer directory `packages/colonizethis_map/lib/src/render/`. Every Dart file under `lib/src/render/` is a PNG-encoding render module permitted to depend on `package:image`; generation passes and view-model builders stay image-free so internal layering (gen → view inputs → render) is one-directional (slice-5 render relocation complete). A same-prefix package such as `package:image_picker/...` does not match; comment lines are ignored (Refs #3574); rule `repo.map_gen_no_image_import` |
 | `tool/check_combat_dedup_precombat_indexes.dart` | Require both combat-phase consumers `combat/conflict_detection.dart` and `combat/unopposed_province_capture.dart` to keep referencing `PreCombatMovementIndex.build` and `greatPowerArmyMoves`, so the deduplicated pre-combat movement/province scan stays delegated to the shared `pre_combat_index.dart` index instead of being re-inlined (Refs #3448); rule `repo.combat_dedup_precombat_indexes` |
-| `tool/check_combat_documented_keep_copies.dart` | Companion to `repo.combat_no_raw_copies_in_resolution`: for combat-resolution copies the #3448 copy-disposition table marks **keep** (`List<Player>.from(...)` in `military_attack_economy.dart`, `copyGroups` `List<String>.from(...)` in `quick_battle_resolver_engine.dart`, `copyWith()).toList()` in `combat_resolver_probabilistic.dart`, `defenderUnitIds.toList()` in `combat_resolver.dart`), require the file to carry an inline `copy-disposition` rationale marker whenever the keep-copy expression is present, so semantically required working-state copies stay documented rather than silently flipped into undocumented clones (Refs #3448); rule `repo.combat_documented_keep_copies` |
+| `tool/check_combat_documented_keep_copies.dart` | Companion to `repo.combat_no_raw_copies_in_resolution`: for combat-resolution copies the #3448 copy-disposition table marks **keep** (`List<Player>.from(...)` in `military_attack_economy.dart`, `copyGroups` `List<String>.from(...)` in `quick_battle_resolver_engine_groups.dart`, `copyWith()).toList()` in `combat_resolver_probabilistic.dart`, `defenderUnitIds.toList()` in `combat_resolver.dart`), require the file to carry an inline `copy-disposition` rationale marker whenever the keep-copy expression is present, so semantically required working-state copies stay documented rather than silently flipped into undocumented clones (Refs #3448); rule `repo.combat_documented_keep_copies` |
 | `tool/check_setup_dedup_init_pipeline_retry.dart` | Forbid re-inlining the init map+setup retry loop under `packages/colonizethis_setup/lib/src/setup/**`: any inline per-attempt seed bump (`* 100003`) must use `kInitPipelineSeedBump` via `runInitPipelineWithRetries`, and the retriable-topology-code predicate (`code == 'faction_component_bin_pack_failed'`) must appear at most once (delegate to `isRetriableInitTopologyCode`); comment lines and throw sites (`code: '...'`) are ignored (Refs #3449); rule `repo.setup_dedup_init_pipeline_retry` |
 | `tool/check_setup_dedup_gp_ow_tile_scans.dart` | Forbid re-inlining the Great Power Old World tile-scan helpers under `packages/colonizethis_setup/lib/src/setup/**` outside the shared `gp_old_world_tile_scan.dart` module: any reappearance of the duplicated private clones (`_ownerByLocalProvinceId`, `_owTileKey`, `_isGpId`, `_collectEligibleTilesSorted`) must instead delegate to the canonical public helpers (`gpOwnerByLocalProvinceId`, `gpOwTileKey`, `isGpOwner`, `visitGpOwLandTiles`, `collectGpOwEligibleTilesSorted`); the shared module is exempt and comment lines are ignored (Refs #3449); rule `repo.setup_dedup_gp_ow_tile_scans` |
 | `tool/check_setup_dedup_gp_ids_from_players.dart` | Forbid re-inlining the slot-ordered Great Power id projection over a player collection (`players.map((p) => p.id)`) anywhere under `packages/colonizethis_setup/lib/src/setup/**` outside the shared `gp_old_world_tile_scan.dart` module: GP id lists/sets must call the canonical `gpIdsSortedFromPlayers(game)` helper (and derive the membership set via `.toSet()`) so player slot order (gp1, gp2, … — never lex-sorted) has one source of truth across the resource/terrain redistribution concerns; the shared module is exempt and comment lines are ignored (Refs #3449); rule `repo.setup_dedup_gp_ids_from_players` |
