+++ packages/colonizethis_combat/lib/src/combat/quick_battle_resolver_engine_strike.dart
@@ -1,0 +1,153 @@
+/// Quick Battle lane effective-strength and strike loss-fraction math.
+///
+/// SPEC/program/quick-battle-resolution.md.
+library;
+
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import 'combat_effective_strength.dart';
+import 'combat_loss_profile.dart';
+import 'quick_battle_action_modifiers.dart';
+import 'quick_battle_emplaced_guns.dart';
+
+double effectiveStrength(
+  List<QuickBattleGroup> groups,
+  Map<String, QuickBattleLaneTerrain> laneTerrain,
+) {
+  var total = 0.0;
+  for (final g in groups) {
+    if (g.unitIds.isEmpty || g.cohesion <= 0) continue;
+    final cohesionScale = g.cohesion / quickBattleMaxCohesion;
+    final terrainKey = '${g.lane.name}_${g.line.name}';
+    final terr = laneTerrain[terrainKey] ?? QuickBattleLaneTerrain.open;
+    var mod = 1.0;
+    switch (terr) {
+      case QuickBattleLaneTerrain.hill:
+      case QuickBattleLaneTerrain.town:
+        mod = 1.1;
+        break;
+      case QuickBattleLaneTerrain.woods:
+        mod = 0.95;
+        break;
+      case QuickBattleLaneTerrain.swamp:
+        mod = 0.8;
+        break;
+      case QuickBattleLaneTerrain.open:
+        break;
+    }
+    total += g.unitIds.length * cohesionScale * mod;
+  }
+  return total;
+}
+
+bool attackerActsFirst(QuickBattleInput input) {
+  final attackerInitiative =
+      input.attackerCavalryShare * initiativeCavalryShareWeight +
+      input.attackerGeneralMedals * initiativeGeneralMedalWeight;
+  final defenderInitiative =
+      input.defenderCavalryShare * initiativeCavalryShareWeight +
+      input.defenderGeneralMedals * initiativeGeneralMedalWeight;
+  if (attackerInitiative != defenderInitiative) {
+    return attackerInitiative > defenderInitiative;
+  }
+  return input.attackerFactionId.compareTo(input.defenderFactionId) < 0;
+}
+
+/// Computes the defender loss fraction from an attacker strike using
+/// precomputed effective strengths.
+///
+/// Callers compute `effAtt` and `effDef` once at the round level via
+/// [attackerEffectiveStrength] and [defenderEffectiveStrength] and only
+/// recompute the side whose state has changed between strikes. This avoids
+/// the previous O(4) [effectiveStrength] invocations per round, where the
+/// striker side's strength was identical across both loss-fraction calls.
+double defenderLossFractionFromAttackerStrike({
+  required QuickBattleInput input,
+  required double effAtt,
+  required double effDef,
+  required ActionModifiers attMods,
+  required ActionModifiers defMods,
+}) {
+  final effAttForRatio = combatEffectiveAttackForRatio(
+    effAtt: effAtt,
+    fortLevel: input.fortLevel,
+  );
+  final ratio = effDef > 0 ? effAttForRatio / effDef : 10.0;
+  return (_targetLossFraction(ratio) *
+          attMods.casualtiesDealtModifier *
+          defMods.casualtiesTakenModifier)
+      .clamp(0.0, 1.0);
+}
+
+/// Computes the attacker loss fraction from a defender strike using
+/// precomputed effective strengths. See [defenderLossFractionFromAttackerStrike]
+/// for the round-level caching contract.
+double attackerLossFractionFromDefenderStrike({
+  required double effAtt,
+  required double effDef,
+  required ActionModifiers attMods,
+  required ActionModifiers defMods,
+}) {
+  final ratio = effAtt > 0 ? effDef / effAtt : 10.0;
+  return (_targetLossFraction(ratio) *
+          defMods.casualtiesDealtModifier *
+          attMods.casualtiesTakenModifier)
+      .clamp(0.0, 1.0);
+}
+
+/// Base loss fraction the striker inflicts on the target, before action and
+/// casualty modifiers.
+///
+/// Classifies the striker-to-target ratio with the shared
+/// [classifyCombatStrengthRatioBand] so the `>= 1.5` / `<= 0.67` band edges stay
+/// the single source of truth shared with the auto-resolve loss profile
+/// (Refs #3448, AC2). The Quick Battle fractions (0.85 / 0.15 / 0.4) remain
+/// distinct from the auto-resolve fractions; only the band thresholds are
+/// shared.
+double _targetLossFraction(double strikerToTargetRatio) {
+  switch (classifyCombatStrengthRatioBand(strikerToTargetRatio)) {
+    case CombatStrengthRatioBand.strongStriker:
+      return 0.85;
+    case CombatStrengthRatioBand.strongTarget:
+      return 0.15;
+    case CombatStrengthRatioBand.even:
+      return 0.4;
+  }
+}
+
+double attackerEffectiveStrength({
+  required QuickBattleInput input,
+  required List<QuickBattleGroup> attGroups,
+  required ActionModifiers attMods,
+}) {
+  final terrainMod = terrainModifiers[input.provinceTerrain] ?? (1.0, 1.0);
+  return combatEffectiveAttackerStrength(
+    base: effectiveStrength(attGroups, input.attackerDeployment.laneTerrain),
+    fortLevel: input.fortLevel,
+    factor1: attMods.offenseModifier,
+    factor2: input.attackerLeaderMultiplier,
+    factor3: terrainMod.$1,
+  );
+}
+
+double defenderEffectiveStrength({
+  required QuickBattleInput input,
+  required List<QuickBattleGroup> defGroups,
+  required ActionModifiers defMods,
+  required List<MutableEmplacedGun> mutableGuns,
+  required bool useVirtualEmplaced,
+}) {
+  final terrainMod = terrainModifiers[input.provinceTerrain] ?? (1.0, 1.0);
+  final emplaced = useVirtualEmplaced
+      ? aliveGunStrengthSum(mutableGuns)
+      : combatDefaultEmplacedStrength(input.fortLevel);
+  return combatEffectiveDefenderStrength(
+    base: effectiveStrength(defGroups, input.defenderDeployment.laneTerrain),
+    fortLevel: input.fortLevel,
+    factor1: defMods.offenseModifier,
+    factor2: input.defenderLeaderMultiplier,
+    factor3: terrainMod.$2,
+    emplacedStrength: emplaced,
+  );
+}
